+++ doc/release-notes.md
@@ -3,3 +3,6 @@ Bitcoin ABC version 0.19.2 is now available from:
   <https://download.bitcoinabc.org/0.19.2/>
 
 This release includes the following features and fixes:
+ - Added parameter `include_removed` to `listsinceblock` for better tracking of
+   transactions during a reorg. See `bitcoin-cli help listsinceblock` for more
+   details.